requirements.txt
streamlit>=1.36
pandas>=2.0
python-dateutil>=2.8
reportlab>=3.6
